# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Anonymous Voting API",
    "description": "API for secure anonymous voting using zk-SNARKs",
    "version": "1.0.0",
    "contact": {
      "email": "admin@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "tags": [
    {
      "name": "auth",
      "description": "Authentication operations"
    },
    {
      "name": "voting",
      "description": "Voting operations"
    },
    {
      "name": "admin",
      "description": "Admin operations"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login with voter ID",
        "description": "Generates zk-SNARK proof for anonymous authentication",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "voterId": {
                    "type": "string",
                    "description": "Voter identifier"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT authentication token"
                    },
                    "nullifierHash": {
                      "type": "string",
                      "description": "Unique hash to prevent double voting"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/api/auth/admin/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Admin login",
        "description": "Login as administrator",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT authentication token"
                    },
                    "isAdmin": {
                      "type": "boolean",
                      "description": "Indicates admin privileges"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/api/auth/verify": {
      "get": {
        "tags": ["auth"],
        "summary": "Verify authentication token",
        "description": "Verifies if the current authentication token is valid",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "isAdmin": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
    "/api/voting/cast": {
      "post": {
        "tags": ["voting"],
        "summary": "Cast a vote",
        "description": "Cast an anonymous vote with zk-SNARK proof",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vote": {
                    "type": "string",
                    "description": "Encrypted vote option"
                  },
                  "nullifierHash": {
                    "type": "string",
                    "description": "Unique hash to prevent double voting"
                  },
                  "proof": {
                    "type": "object",
                    "description": "zk-SNARK proof"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vote cast successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "transactionHash": {
                      "type": "string",
                      "description": "Blockchain transaction hash"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or proof"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Already voted"
          }
        }
      }
    },
    "/api/voting/status": {
      "get": {
        "tags": ["voting"],
        "summary": "Get voting status",
        "description": "Check if a voter has already cast a vote",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nullifierHash",
            "schema": {
              "type": "string"
            },
            "description": "Nullifier hash of the voter"
          }
        ],
        "responses": {
          "200": {
            "description": "Voting status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hasVoted": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/logs": {
      "get": {
        "tags": ["admin"],
        "summary": "Get system logs",
        "description": "Retrieve system logs for monitoring",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "timeFrame",
            "schema": {
              "type": "string",
              "enum": ["1h", "24h", "7d", "30d"]
            },
            "description": "Time frame for logs"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Number of logs per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "timestamp": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "level": {
                            "type": "string",
                            "enum": ["info", "warning", "error"]
                          },
                          "message": {
                            "type": "string"
                          },
                          "meta": {
                            "type": "object"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        },
                        "current": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    },
    "/api/admin/status": {
      "get": {
        "tags": ["admin"],
        "summary": "Get system status",
        "description": "Retrieve system statistics and status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "System status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalVotes": {
                      "type": "integer"
                    },
                    "uptime": {
                      "type": "number",
                      "description": "System uptime in seconds"
                    },
                    "blockchainStatus": {
                      "type": "string",
                      "enum": ["connected", "disconnected", "syncing"]
                    },
                    "databaseStatus": {
                      "type": "string",
                      "enum": ["connected", "disconnected"]
                    },
                    "voteDistribution": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer"
                      },
                      "description": "Distribution of votes by option"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 